+++ src/io/react-native/react-native-worker.ts
@@ -17,7 +17,7 @@ import {
   EdgeFetchResponse,
   EdgeIo
 } from '../../types/types'
-import { initMixFetch, mixFetchOptions } from '../../util/nym'
+import { nymFetch } from '../../util/nym'
 import { hideProperties } from '../hidden-properties'
 import { makeNativeBridge } from './native-bridge'
 import { WorkerApi, YAOB_THROTTLE_MS } from './react-native-types'
@@ -176,16 +176,7 @@ async function makeIo(logBackend: LogBackend): Promise<EdgeIo> {
       const { corsBypass = 'auto', privacy = 'none' } = opts ?? {}
 
       if (privacy === 'nym') {
-        const nymFetch = await initMixFetch(log)
-        const response = await nymFetch(
-          uri,
-          {
-            ...opts,
-            mode: 'unsafe-ignore-cors' as RequestMode
-          },
-          mixFetchOptions
-        )
-        return response
+        return await nymFetch(uri, opts ?? {}, log)
       }
       if (corsBypass === 'always') {
         return await nativeFetch(uri, opts)